Move ata commands to folder
